podcast-automation/requirements.txt
streamlit
google-cloud-texttospeech
pydub
requests
python-dotenv
boto3
vertexai
elevenlabs
openai